# internal/transport/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: HCAI CHAT API
  version: 0.2.0
  description: Contract for the international-first HCAI CHAT platform.
servers:
  - url: /api/v1
...
paths:
...
  /principal:
    get:
      tags: [Developer]
      operationId: getDeveloperAPIPrincipal
      security: [{ bearerApiKey: [] }]
      responses:
        '200': { description: Authenticated Service Account principal envelope }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /errors:
    get:
      tags: [Developer]
      operationId: getDeveloperAPIErrorRegistry
      security: [{ bearerApiKey: [] }]
      responses:
        '200': { description: Stable Developer API v1 error registry envelope }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /payments/webhooks/stripe:
    post:
      tags: [Payments]
      operationId: receiveStripeWebhook
      description: Receives a bounded, signed Stripe event. The endpoint is disabled unless the payment Provider is explicitly configured.
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema: { type: string, maxLength: 4096 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Duplicate event safely acknowledged, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentWebhookReceipt' } } } }
        '202': { description: Signed event accepted for idempotent processing, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentWebhookReceipt' } } } }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/ValidationFailed' }
        '415': { $ref: '#/components/responses/ValidationFailed' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /auth/session:
    get:
      tags: [Identity]
      operationId: getSession
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      tags: [Identity]
      operationId: registerAccount
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } }
      responses:
        '201':
          description: Account and hashed-cookie session created
          content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /auth/login:
    post:
      tags: [Identity]
      operationId: loginAccount
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200':
          description: Email and password session created
          content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/demo:
    post:
      tags: [Identity]
      operationId: startDemoSession
      description: Starts a clearly labeled local-development session. The endpoint is unavailable in production.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actor:
                  type: string
                  enum: [creator, publisher, admin]
      responses:
        '200':
          description: Demo session established
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /auth/password-reset-requests:
    post:
      tags: [Identity]
      operationId: requestPasswordReset
      description: Always accepts the request without revealing whether the email belongs to an active account.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } }
      responses:
        '202': { description: Request accepted without an account-existence signal }
  /auth/password-reset-confirm:
    post:
      tags: [Identity]
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetConfirm' } } }
      responses:
        '200': { description: Password replaced, all sessions revoked, and the one-time token erased }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { description: Email action expired }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /auth/email-verification/confirm:
    post:
      tags: [Identity]
      operationId: confirmEmailVerification
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EmailActionToken' } } }
      responses:
        '204': { description: Email verified and one-time token erased }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { description: Email action expired }
  /account/profile:
    patch:
      tags: [Identity]
      operationId: updateProfile
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProfileUpdate' } } }
      responses:
        '200':
          description: Current profile updated with audit evidence
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties: { user: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /account/payouts:
    get:
      tags: [Payments]
      operationId: getPayoutStatus
      security: [{ sessionCookie: [] }]
      description: Returns minimized owner-scoped Stripe Connect payout onboarding evidence. Identity, bank, and requirement details are never returned.
      responses:
        '200':
          description: Current creator payout onboarding state
          content: { application/json: { schema: { $ref: '#/components/schemas/PayoutStatus' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /account/payouts/onboarding:
    post:
      tags: [Payments]
      operationId: beginPayoutOnboarding
      security: [{ sessionCookie: [] }]
      description: Creates a short-lived, single-use Stripe-hosted Connect onboarding link for the authenticated creator.
      responses:
        '201':
          description: Hosted onboarding link created; the URL is returned only to the authenticated owner.
          content: { application/json: { schema: { $ref: '#/components/schemas/PayoutOnboardingLink' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /account/email-actions:
    get:
      tags: [Identity]
      operationId: listAccountEmailActions
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 5 } }
        - { name: cursor, in: query, schema: { type: string }, description: Opaque stable owner identity-email-action cursor }
      responses:
        '200':
          description: Safe delivery and lifecycle evidence for the current account; token material is never returned.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: '#/components/schemas/IdentityEmailAction' } }, nextCursor: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /account/email-verification:
    post:
      tags: [Identity]
      operationId: requestEmailVerification
      security: [{ sessionCookie: [] }]
      responses:
        '202': { description: Durable verification action queued, content: { application/json: { schema: { $ref: '#/components/schemas/IdentityEmailAction' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Email is already verified }
  /account/sessions:
    get:
      tags: [Identity]
      operationId: listAccountSessions
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: cursor, in: query, schema: { type: string }, description: Opaque stable account-session cursor }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: Current account sessions with safe device and network hints
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountSessionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /account/sessions/{sessionId}:
    delete:
      tags: [Identity]
      operationId: revokeAccountSession
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        '204': { description: Owned session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /account/sessions/revoke-others:
    post:
      tags: [Identity]
      operationId: revokeOtherAccountSessions
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: All other active sessions revoked
          content:
            application/json:
              schema:
                type: object
                required: [revokedCount]
                properties: { revokedCount: { type: integer, minimum: 0 } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /account/developer-access:
    get:
      tags: [Developer]
      operationId: getDeveloperAccess
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Personal control state and safe credential inventory, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperAccess' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /account/developer-service-accounts:
    post:
      tags: [Developer]
      operationId: createDeveloperServiceAccount
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperServiceAccountCreate' } } } }
      responses:
        '201': { description: Service Account created, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperServiceAccount' } } } }
        '503': { description: Developer Access is disabled }
  /account/developer-service-accounts/{accountId}/revoke:
    post:
      tags: [Developer]
      operationId: revokeDeveloperServiceAccount
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/DeveloperAccountId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperTransition' } } } }
      responses:
        '200': { description: Service Account and active keys revoked, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperServiceAccount' } } } }
  /account/developer-service-accounts/{accountId}/keys:
    post:
      tags: [Developer]
      operationId: issueDeveloperAPIKey
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/DeveloperAccountId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperKeyCreate' } } } }
      responses:
        '201': { description: API key issued with one-time plaintext, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperCredential' } } } }
  /account/developer-service-accounts/{accountId}/keys/{keyId}/rotate:
    post:
      tags: [Developer]
      operationId: rotateDeveloperAPIKey
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/DeveloperAccountId' }, { $ref: '#/components/parameters/DeveloperKeyId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperKeyRotate' } } } }
      responses:
        '201': { description: Replacement API key issued and predecessor invalidated, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperCredential' } } } }
  /account/developer-service-accounts/{accountId}/keys/{keyId}/revoke:
    post:
      tags: [Developer]
      operationId: revokeDeveloperAPIKey
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/DeveloperAccountId' }, { $ref: '#/components/parameters/DeveloperKeyId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperTransition' } } } }
      responses:
        '200': { description: API key revoked, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperAPIKey' } } } }
  /account/developer-webhooks:
    get:
      tags: [Developer]
      operationId: getDeveloperWebhooks
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Owned webhook endpoints, event catalog, and delivery evidence without signing secrets, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperWebhookAccess' } } } }
    post:
      tags: [Developer]
      operationId: createDeveloperWebhook
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperWebhookCreate' } } } }
      responses:
        '201': { description: Webhook endpoint created with one-time plaintext signing secret, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperWebhookCredential' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { description: Developer Access is disabled }
...
  /account/developer-webhooks/{endpointId}/rotate:
    post:
      tags: [Developer]
      operationId: rotateDeveloperWebhookSecret
      security: [{ sessionCookie: [] }]
      parameters: [{ name: endpointId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperTransition' } } } }
      responses:
        '201': { description: New signing-secret revision created and returned once, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperWebhookCredential' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /account/developer-webhooks/{endpointId}/revoke:
    post:
      tags: [Developer]
      operationId: revokeDeveloperWebhook
      security: [{ sessionCookie: [] }]
      parameters: [{ name: endpointId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperTransition' } } } }
      responses:
        '200': { description: Endpoint revoked and pending deliveries cancelled, content: { application/json: { schema: { $ref: '#/components/schemas/DeveloperWebhookEndpoint' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /auth/oauth/{provider}/start:
    post:
      tags: [Identity]
      operationId: startOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, github] }
      responses:
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hcai_session
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: hcai_sk
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: Required provider unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    PaymentWebhookReceipt:
      type: object
      required: [eventId, providerEventId, eventType, status, duplicate]
      properties:
        eventId: { type: string, format: uuid }
        providerEventId: { type: string }
        eventType: { type: string }
        status: { type: string, enum: [received, processing, retry_scheduled, processed, ignored, failed] }
        duplicate: { type: boolean }
...